build: consolidate CMake presets, fix warnings, pin clang-format 20
- Consolidate multi-file CMake presets into single CMakePresets.json
  with shorter names (cl-debug, clang-debug, gcc-debug, etc.)
- Fix clang-cl warning handling for Windows builds
- Fix scan512 COMPILE_FLAGS overwrite in CMake
- Improve third-party warning suppression
- Suppress unused CMAKE_TOOLCHAIN_FILE warning
- Fix sign-conversion warnings in SIMD code
- Suppress Windows GUI dialogs during test execution
- Update CI workflow preset names
- Pin CI clang-format to version 20

+++ src/libunicode/simd_detector.cpp
@@ -12,7 +12,7 @@
 
 void cpuid(int32_t out[4], int32_t eax, int32_t ecx);
 #if _WIN32
-__int64 xgetbv(unsigned int x);
+uint64_t xgetbv(unsigned int x);
 #elif defined(__GNUC__) || defined(__clang__)
 uint64_t xgetbv(unsigned int index);
 #else
@@ -30,9 +30,9 @@ auto detect_os_avx512() -> bool;
         {
             __cpuidex(out, eax, ecx);
         }
-        __int64 xgetbv(unsigned int x)
+        uint64_t xgetbv(unsigned int x)
         {
-            return _xgetbv(x);
+            return static_cast<uint64_t>(_xgetbv(x));
         }
     // clang-format on
     #elif defined(__GNUC__) || defined(__clang__)
